Move contracts into builtin-contracts.java, rename test
